+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 - **fix**: create and import problem using keyboard shortcut
 - **feat**: focus back on editor after keyboard shortcut
 - **feat**: changed default value of cph-ng.comparing.oleSize to 8
+- **feat**: let llm decide active file path
 
 ## 0.4.6